Update Monster.py
added the game itself to the play card thing, as it is needed for the AIs (to make shallow game copies)
Am going to add a way to make a virtual copy of the game. Moved the game class up just to make sure it doesn't break when players reference creating games. Added a "quiet mode" to the game class.
Oh, and split up playing a hand from a full game, for ease of AI testing.
please ask any further questions in the discord.